requirements.txt
Werkzeug==0.14.1
moviepy==0.2.3.5
scipy==1.1.0
tensorflow==1.10.1
numpy==1.15.1
Pillow==5.2.0
Flask==1.0.2
